+++ app/models/__init__.py
@@ -1,3 +1,3 @@
-from .linguistic import WordToken, ChangeRecord, Corpus, AllowedLemma, AllowedMorph, AllowedPOS, \
-    ControlListsUser, ControlLists, CorpusUser
+from .corpus import WordToken, ChangeRecord, Corpus, CorpusUser
 from .user import User, AnonymousUser, Permission, Role
+from .control_lists import AllowedLemma, AllowedMorph, AllowedPOS, ControlListsUser, ControlLists
