+++ tests/test_models/test_allowed_morph.py
@@ -1,5 +1,5 @@
 from .base import TestModels
-from app.models.linguistic import AllowedMorph
+from app.models import AllowedMorph
 
 
 class TestAllowedMorph(TestModels):
